requirements_audio.txt
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
imutils==0.5.2
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
kiwisolver==1.0.1
mahotas==1.4.5
matplotlib==3.0.3
networkx==2.3
numpy==1.16.2
opencv-contrib-python==3.4.4.19
pandas==0.24.2
parso==0.3.4
pexpect==4.6.0
picamera==1.13
pickleshare==0.7.5
pillow>=6.2.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
PyAudio==0.2.11
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
traitlets==4.3.2
Wave==0.0.2
wcwidth==0.1.7
